# apidog/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/purchase_orders": {
      "get": {
        "operationId": "PurchaseController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Text search across string columns",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "description": "Return soft-deleted records too (true/false)",
            "schema": {
              "default": false,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "confirmed",
                "received",
                "partial",
                "cancelled"
              ]
            }
          },
          {
            "name": "supplier_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List purchase orders",
        "tags": [
          "PurchaseOrders"
        ]
      },
      "post": {
        "operationId": "PurchaseController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create purchase order with items",
        "tags": [
          "PurchaseOrders"
        ]
      }
    },
    "/api/v1/purchase_orders/{id}": {
      "get": {
        "operationId": "PurchaseController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get purchase order details by id",
        "tags": [
          "PurchaseOrders"
        ]
      },
      "patch": {
        "operationId": "PurchaseController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePurchaseOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update purchase order header by id",
        "tags": [
          "PurchaseOrders"
        ]
      }
    },
    "/api/v1/purchase_orders/{id}/items": {
      "post": {
        "operationId": "PurchaseController_addItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPurchaseOrderItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Add item to purchase order",
        "tags": [
          "PurchaseOrders"
        ]
      }
    },
    "/api/v1/purchase_orders/{id}/items/{itemId}": {
      "patch": {
        "operationId": "PurchaseController_updateItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePurchaseOrderItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update purchase order item",
        "tags": [
          "PurchaseOrders"
        ]
      },
      "delete": {
        "operationId": "PurchaseController_removeItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete purchase order item",
        "tags": [
          "PurchaseOrders"
        ]
      }
    },
    "/api/v1/purchase_return": {
      "get": {
        "operationId": "PurchaseReturnController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Text search across string columns",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "description": "Return soft-deleted records too (true/false)",
            "schema": {
              "default": false,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "completed",
                "cancelled"
              ]
            }
          },
          {
            "name": "supplier_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "purchase_order_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List purchase returns",
        "tags": [
          "PurchaseReturns"
        ]
      },
      "post": {
        "operationId": "PurchaseReturnController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseReturnDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create purchase return (purchase_order_id in payload) and deduct stock",
        "tags": [
          "PurchaseReturns"
        ]
      }
    },
    "/api/v1/purchase_return/{id}": {
      "get": {
        "operationId": "PurchaseReturnController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get purchase return details by id",
        "tags": [
          "PurchaseReturns"
        ]
      },
      "patch": {
        "operationId": "PurchaseReturnController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePurchaseReturnDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update purchase return; re-applies stock when items change",
        "tags": [
          "PurchaseReturns"
        ]
      },
      "delete": {
        "operationId": "PurchaseReturnController_softDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Soft delete purchase return and reverse stock",
        "tags": [
          "PurchaseReturns"
        ]
      }
    },
    "/api/v1/purchase_return/{id}/permanent": {
      "delete": {
        "operationId": "PurchaseReturnController_permanentDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Permanently delete purchase return and reverse stock",
        "tags": [
          "PurchaseReturns"
        ]
      }
    },
...
  },
  "info": {
    "title": "Pharmacy Management API",
    "description": "Pharmacy backend API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "description": "Enter JWT access token"
      }
    },
    "schemas": {
...
      "CreatePurchaseOrderItemDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid"
          },
          "product_batch_id": {
            "type": "string",
            "format": "uuid"
          },
          "purchase_unit_id": {
            "type": "string",
            "format": "uuid"
          },
          "quantity_purchase": {
            "type": "number",
            "description": "Quantity in purchase unit"
          },
          "unit_cost": {
            "type": "number",
            "description": "Cost per purchase unit"
          },
          "discount": {
            "type": "number"
          },
          "tax": {
            "type": "number"
          },
          "expected_expiry_date": {
            "type": "string",
            "format": "date"
          },
          "batch_number": {
            "type": "string",
            "maxLength": 80
          }
        },
        "required": [
          "product_id",
          "quantity_purchase",
          "unit_cost"
        ]
      },
      "CreatePurchaseOrderDto": {
        "type": "object",
        "properties": {
          "po_number": {
            "type": "string",
            "maxLength": 80
          },
          "supplier_id": {
            "type": "string",
            "format": "uuid"
          },
          "shop_id": {
            "type": "string",
            "format": "uuid"
          },
          "branch_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed",
              "received",
              "partial",
              "cancelled"
            ]
          },
          "phar_payment_status": {
            "type": "string",
            "enum": [
              "pending",
              "partial",
              "paid",
              "failed",
              "refunded",
              "cancelled"
            ]
          },
          "discount_amount": {
            "type": "number"
          },
          "tax_amount": {
            "type": "number"
          },
          "shipping_cost": {
            "type": "number"
          },
          "expected_delivery_date": {
            "type": "string",
            "format": "date"
          },
          "delivery_date": {
            "type": "string",
            "format": "date"
          },
          "notes": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePurchaseOrderItemDto"
            }
          }
        },
        "required": [
          "supplier_id",
          "items"
        ]
      },
      "UpdatePurchaseOrderDto": {
        "type": "object",
        "properties": {
          "po_number": {
            "type": "string",
            "maxLength": 80
          },
          "supplier_id": {
            "type": "string",
            "format": "uuid"
          },
          "shop_id": {
            "type": "string",
            "format": "uuid"
          },
          "branch_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed",
              "received",
              "partial",
              "cancelled"
            ]
          },
          "phar_payment_status": {
            "type": "string",
            "enum": [
              "pending",
              "partial",
              "paid",
              "failed",
              "refunded",
              "cancelled"
            ]
          },
          "discount_amount": {
            "type": "number"
          },
          "tax_amount": {
            "type": "number"
          },
          "shipping_cost": {
            "type": "number"
          },
          "expected_delivery_date": {
            "type": "string",
            "format": "date"
          },
          "delivery_date": {
            "type": "string",
            "format": "date"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "AddPurchaseOrderItemDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid"
          },
          "product_batch_id": {
            "type": "string",
            "format": "uuid"
          },
          "purchase_unit_id": {
            "type": "string",
            "format": "uuid"
          },
          "quantity_purchase": {
            "type": "number",
            "description": "Quantity in purchase unit"
          },
          "unit_cost": {
            "type": "number",
            "description": "Cost per purchase unit"
          },
          "discount": {
            "type": "number"
          },
          "tax": {
            "type": "number"
          },
          "expected_expiry_date": {
            "type": "string",
            "format": "date"
          },
          "batch_number": {
            "type": "string",
            "maxLength": 80
          }
        },
        "required": [
          "product_id",
          "quantity_purchase",
          "unit_cost"
        ]
      },
      "UpdatePurchaseOrderItemDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid"
          },
          "product_batch_id": {
            "type": "string",
            "format": "uuid"
          },
          "purchase_unit_id": {
            "type": "string",
            "format": "uuid"
          },
          "quantity_purchase": {
            "type": "number",
            "description": "Quantity in purchase unit"
          },
          "unit_cost": {
            "type": "number",
            "description": "Cost per purchase unit"
          },
          "discount": {
            "type": "number"
          },
          "tax": {
            "type": "number"
          },
          "expected_expiry_date": {
            "type": "string",
            "format": "date"
          },
          "batch_number": {
            "type": "string",
            "maxLength": 80
          }
        }
      },
      "CreatePurchaseReturnItemDto": {
        "type": "object",
        "properties": {
          "purchase_order_item_id": {
            "type": "string",
            "format": "uuid"
          },
          "product_batch_id": {
            "type": "string",
            "format": "uuid"
          },
          "return_unit_id": {
            "type": "string",
            "format": "uuid"
          },
          "return_qty": {
            "type": "number",
            "description": "Return quantity in return unit"
          },
          "unit_cost": {
            "type": "number",
            "description": "Cost per return unit override"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "purchase_order_item_id",
          "return_qty"
        ]
      },
      "CreatePurchaseReturnDto": {
        "type": "object",
        "properties": {
          "purchase_order_id": {
            "type": "string",
            "format": "uuid",
            "description": "Purchase order the return is created against"
          },
          "return_number": {
            "type": "string",
            "maxLength": 80
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "completed",
              "cancelled"
            ]
          },
          "reason": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePurchaseReturnItemDto"
            }
          }
        },
        "required": [
          "purchase_order_id",
          "items"
        ]
      },
      "UpdatePurchaseReturnDto": {
        "type": "object",
        "properties": {
          "return_number": {
            "type": "string",
            "maxLength": 80
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "completed",
              "cancelled"
            ]
          },
          "reason": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "items": {
            "description": "When provided, fully replaces the return lines. Existing stock effects are reversed and the new lines are re-applied.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePurchaseReturnItemDto"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}